ref(replay): Extract replay sending into functions
